Add multipart body part documentation
